+++ test/fixtures/classes.js
@@ -21,10 +21,24 @@ class Rufus extends Error {
 	}
 }
 
+class Rosie extends Error {
+	constructor(Bentley, Olive) {
+		super(Bentley, Olive);
+		console.log(Bentley, Olive);
+	}
+}
+
 class Josie extends Error {
 	constructor(Daisy = 'Daisy') {
 		super(Daisy);
 		console.log(Daisy);
+	}
+}
+
+class Lady extends Error {
+	constructor(Chase = 'Chase', Izzy) {
+		super(Chase, Izzy);
+		console.log(Chase, Izzy);
 	}
 }
 
@@ -34,5 +48,12 @@ class Hannah extends Error {
 	constructor(Tucker = 'Tucker') {
 		super(Charlie);
 		console.log(Tucker);
+	}
+}
+
+class Oreo extends Error {
+	constructor(Max = 'Max', Molly = 'Molly') {
+		super(Rex, Baxter);
+		console.log(Max, Molly);
 	}
 }